Extract covet to internal packages, prepare for sharing summary report funcs
